--- packages.json
@@ -1,7 +1,0 @@
-{
-  "dependencies": {
-    "next": "^12.0.0", // or the latest version
-    "react": "^17.0.0",
-    "react-dom": "^17.0.0"
-  }
-}
